backend/requirements_test.txt
django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
gunicorn==20.1.0
django-rosetta==0.9.9
polib==1.2.0
charset-normalizer==3.4.4
idna==3.11